% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRball-package.R
\docType{package}
\name{CRISPRball-package}
\alias{CRISPRball-package}
\title{CRISPRball: Shiny Application for Interactive CRISPR Screen Visualization, Exploration, Comparison, and Filtering}
\description{
A Shiny application for visualization, exploration, comparison, and filtering of CRISPR screens analyzed with MAGeCK RRA or MLE. Features include interactive plots with on-click labeling, full customization of plot aesthetics, data upload and/or download, and much more. Quickly and easily explore your CRISPR screen results and generate publication-quality figures in seconds.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/j-andrews7/CRISPRball}
  \item Report bugs at \url{https://support.bioconductor.org/}
}

}
\author{
\strong{Maintainer}: Jared Andrews \email{jared.andrews07@gmail.com} (\href{https://orcid.org/0000-0002-0780-6248}{ORCID})

Other contributors:
\itemize{
  \item Jacob Steele (\href{https://orcid.org/0000-0001-9924-2226}{ORCID}) [contributor]
}

}
\keyword{internal}
