% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depmap_utils.R
\name{.make_gene_tag}
\alias{.make_gene_tag}
\title{Generate gene tagList via mygene API}
\usage{
.make_gene_tag(gene)
}
\arguments{
\item{gene}{Character scalar for gene symbol.}
}
\value{
TagList containing dependency summary information.
}
\description{
Generate gene tagList via mygene API
}
