% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depmap_utils.R
\name{get_depmap_essentiality}
\alias{get_depmap_essentiality}
\title{Get essential/selective gene information from DepMap summary table.}
\usage{
get_depmap_essentiality(gene, depmap.summary)
}
\arguments{
\item{gene}{Character scalar for gene symbol.}

\item{depmap.summary}{data.frame containing DepMap summary information.}
}
\value{
Named list containing RNAi and CRISPR named lists containing dataset information for the provided gene,
if available. If the gene is not found in the summary data.frame, the \code{avail} element for the RNAi and CRISPR lists
will be set to \code{FALSE}.
}
\description{
Get essential/selective gene information from DepMap summary table.
}
\author{
Jared Andrews
}
