% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_comparison.R
\name{.create_tab_comparison}
\alias{.create_tab_comparison}
\title{Create a tabPanel for the comparison tab}
\usage{
.create_tab_comparison(datasets)
}
\arguments{
\item{datasets}{A character vector containing dataset names.}
}
\value{
A \code{\link{tabPanel}} with UI elements for the comparison tab.
}
\description{
Create a \code{\link{tabPanel}} with UI elements for the comparison tab.
}
\author{
Jared Andrews
}
