% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_sgrna.R
\name{.create_tab_sgrna}
\alias{.create_tab_sgrna}
\title{Create a tabPanel for the sgrna tab}
\usage{
.create_tab_sgrna(sgrna.choices, sgrna.gene)
}
\arguments{
\item{sgrna.choices}{A character vector containing dataset names.}

\item{sgrna.gene}{A character vector containing gene identifiers.}
}
\value{
A \code{\link{tabPanel}} with UI elements for the sgrna tab.
}
\description{
Create a \code{\link{tabPanel}} with UI elements for the sgrna tab.
}
\author{
Jared Andrews
}
