% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observers_upload.R
\name{.create_upload_observers}
\alias{.create_upload_observers}
\title{Define observers for data upload}
\usage{
.create_upload_observers(input, session, robjects)
}
\arguments{
\item{input}{The Shiny input object from the server function.}

\item{session}{The Shiny session object from the server function.}

\item{robjects}{A reactive list of values generated in the \code{\link{iSEE}} app.}
}
\value{
Observers are set up to monitor the data upload UI elements that can accept new data.
A \code{NULL} is invisibly returned.
}
\description{
Define a series of observers to track additional new data being uploaded.
}
\author{
Jared Andrews
}
