% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_depmap.R
\name{plot_depmap_cn}
\alias{plot_depmap_cn}
\title{Plot gene copy number information from DepMap}
\usage{
plot_depmap_cn(df, color = "#CEA3CB", plot.grid = FALSE)
}
\arguments{
\item{df}{data.frame containing information for a single gene
as returned by \code{\link{get_depmap_plot_data}}.}

\item{color}{Character scalar for trace color.}

\item{plot.grid}{Boolean indicating whether to plot gridlines.}
}
\value{
plotly object
}
\description{
Plot gene copy number information from DepMap
}
\examples{
library(CRISPRball)
data(depmap_22q1_cn)
plot_depmap_cn(depmap_22q1_cn)
}
\seealso{
\code{\link{get_depmap_plot_data}}
}
\author{
Jared Andrews
}
