% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_depmap_lineages}
\alias{plot_depmap_lineages}
\title{Plot selected information across lineages from DepMap.}
\usage{
plot_depmap_lineages(
  gene,
  data.type,
  group.by,
  depmap.meta,
  depmap.pool,
  jitter = 0
)
}
\arguments{
\item{gene}{Character scalar for gene symbol.}

\item{data.type}{One of "crispr", "rnai", "cn", or "ccle_tpm".}

\item{depmap.meta}{data.frame of DepMap cell line metadata, as stored in the 'meta' table
of the SQLite database built by \code{\link{build_depmap_db}}.}

\item{depmap.pool}{pool connection to DepMap SQLite database built with \code{\link{build_depmap_db}}.}
}
\value{
plotly object
}
\description{
Plot selected information across lineages from DepMap.
}
\author{
Jared Andrews
}
