% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_qc.R
\name{plot_pca_biplot}
\alias{plot_pca_biplot}
\title{Plot a biplot from a PCAtools \code{pca} object}
\usage{
plot_pca_biplot(
  pca.res,
  dim.x = "PC1",
  dim.y = "PC2",
  dim.z = NULL,
  plot.title = "PCA Biplot",
  color.by = NULL,
  shape.by = NULL,
  hover.info = NULL,
  show.loadings = FALSE,
  n.loadings = 3,
  pt.size = 12
)
}
\arguments{
\item{pca.res}{A \code{\link[PCAtools]{pca}} generated by the PCAtools package}

\item{dim.x}{The principal component to plot on the x-axis (default "PC1")}

\item{dim.y}{The principal component to plot on the y-axis (default "PC2")}

\item{dim.z}{The principal component to plot on the z-axis (default NULL)}

\item{plot.title}{The title of the plot (default "PCA Biplot")}

\item{color.by}{Column name to use for coloring points (default NULL)}

\item{shape.by}{Column name to use for shaping points (default NULL)}

\item{hover.info}{Column name(s) to include in the hover info for each point (default NULL)}

\item{show.loadings}{Whether to show component loadings on the plot (default FALSE)}

\item{n.loadings}{The number of loadings to show (default 3)}

\item{pt.size}{The size of the plotted points (default 12)}
}
\value{
A plotly plot of the PCA biplot
}
\description{
This function plots a biplot from a PCAtools \code{\link[PCAtools]{pca}} object.
}
\examples{
library("PCAtools")
library("CRISPRball")
col <- 10
row <- 2000
mat <- matrix(
    rexp(col * row, rate = 0.1),
    ncol = col
)
rownames(mat) <- paste0("gene", 1:nrow(mat))
colnames(mat) <- paste0("sample", 1:ncol(mat))

metadata <- data.frame(row.names = colnames(mat))
metadata$Group <- rep(NA, ncol(mat))
metadata$Group[seq(1, 10, 2)] <- "A"
metadata$Group[seq(2, 10, 2)] <- "B"

p <- pca(mat, metadata = metadata, removeVar = 0.1)
plot_pca_biplot(p, color.by = "Group")
}
\seealso{
\code{\link[PCAtools]{pca}}
}
\author{
Jared Andrews
}
