% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_depmap.R
\name{build_depmap_db}
\alias{build_depmap_db}
\title{Build SQLite database of DepMap data}
\usage{
build_depmap_db()
}
\value{
Name of SQLite database containing DepMap data.
}
\description{
Build SQLite database of DepMap data
}
\examples{
\dontrun{
build_depmap_db()
}
}
\author{
Jared Andrews
}
