% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_depmap.R
\name{build_depmap_db}
\alias{build_depmap_db}
\title{Build SQLite database of DepMap data}
\usage{
build_depmap_db(file = "depmap_db.sqlite")
}
\arguments{
\item{file}{Name of SQLite database file to create.}
}
\value{
Name of SQLite database containing DepMap data.
}
\description{
Build SQLite database of DepMap data
}
\examples{
library(CRISPRball)
build_depmap_db()
}
\seealso{
\code{\link[depmap]{depmap_rnai}}, \code{\link[depmap]{depmap_crispr}},
\code{\link[depmap]{depmap_copyNumber}}, \code{\link[depmap]{depmap_TPM}},
\code{\link[depmap]{depmap_metadata}}, \code{\link[depmap]{depmap_gene_summary}},
\code{\link[pool]{dbPool}}, \code{\link[DBI]{dbWriteTable}}
}
\author{
Jared Andrews
}
