% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depmap_utils.R
\name{.make_dependency_tag}
\alias{.make_dependency_tag}
\title{Generate dependency summary info tagList}
\usage{
.make_dependency_tag(dep.info, dep.release)
}
\arguments{
\item{dep.info}{Named list containing summary CRISPR and RNAi info.}

\item{dep.release}{Character scalar for DepMap release as returned by \code{\link[depmap]{depmap_release}}.}
}
\description{
Generate dependency summary info tagList
}
