% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_depmap_cn}
\alias{plot_depmap_cn}
\title{Plot gene copy number information from DepMap, mostly from CCLE}
\usage{
plot_depmap_cn(gene, depmap.meta, depmap.pool, color, plot.grid)
}
\arguments{
\item{gene}{Character scalar for gene symbol.}

\item{depmap.meta}{data.frame of DepMap cell line metadata, as stored in the 'meta' table
of the SQLite database built by \code{\link{build_depmap_db}}.}

\item{depmap.pool}{pool connection to DepMap SQLite database built with \code{\link{build_depmap_db}}.}

\item{color}{Character scalar for trace color.}

\item{plot.grid}{Boolean indicating whether to plot gridlines.}
}
\value{
plotly object
}
\description{
Plot gene copy number information from DepMap, mostly from CCLE
}
\author{
Jared Andrews
}
