% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_depmap_dependency}
\alias{plot_depmap_dependency}
\title{Plot gene dependency information from DepMap CRISPR and RNAi tables.}
\usage{
plot_depmap_dependency(
  gene,
  depmap.meta,
  crispr.color,
  rnai.color,
  depline,
  depmap.pool,
  plot.grid
)
}
\arguments{
\item{gene}{Character scalar for gene symbol.}

\item{depmap.meta}{data.frame of DepMap cell line metadata, as stored in the 'meta' table
of the SQLite database built by \code{\link{build_depmap_db}}.}

\item{crispr.color}{Character scalar for CRISPR trace color.}

\item{rnai.color}{Character scalar for RNAi trace color.}

\item{depline}{Boolean indicating whether to show the dependency line.}

\item{depmap.pool}{pool connection to DepMap SQLite database built with \code{\link{build_depmap_db}}.}

\item{plot.grid}{Boolean indicating whether to plot gridlines.}
}
\value{
plotly object
}
\description{
Plot gene dependency information from DepMap CRISPR and RNAi tables.
}
\author{
Jared Andrews
}
