% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depmap_utils.R
\name{plot_depmap_dependency}
\alias{plot_depmap_dependency}
\title{Plot gene dependency information from DepMap CRISPR and RNAi tables.}
\usage{
plot_depmap_dependency(gene, depmap.meta, depmap.pool)
}
\arguments{
\item{gene}{Character scalar for gene symbol.}

\item{depmap.meta}{data.frame of DepMap cell line metadata, as stored in the 'meta' table
of the SQLite database built by \code{\link{build_depmap_db}}.}

\item{depmap.pool}{pool connection to DepMap SQLite database built with \code{\link{build_depmap_db}}.}
}
\value{
plotly object
}
\description{
Plot gene dependency information from DepMap CRISPR and RNAi tables.
}
\author{
Jared Andrews
}
