% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_depmap_lineages}
\alias{plot_depmap_lineages}
\title{Plot selected information across lineages from DepMap.}
\usage{
plot_depmap_lineages(
  gene,
  data.type,
  group.by,
  depmap.meta,
  depmap.pool,
  lineage = NULL,
  depline = TRUE,
  label.size = 12,
  pt.size = 5,
  pt.color = "#56B4E9",
  boxplot.fill = "#E2E2E2",
  boxplot.line.color = "#000000"
)
}
\arguments{
\item{gene}{Character scalar for gene symbol.}

\item{data.type}{One of "crispr", "rnai", "cn", or "ccle_tpm".}

\item{group.by}{Character scalar of column name to group by.}

\item{depmap.meta}{data.frame of DepMap cell line metadata, as stored in the 'meta' table
of the SQLite database built by \code{\link{build_depmap_db}}.}

\item{depmap.pool}{pool connection to DepMap SQLite database built with \code{\link{build_depmap_db}}.}

\item{lineage}{Character scalar of lineage for which to plot sub-lineage data.}

\item{depline}{Boolean indicating whether to show the dependency line.}

\item{label.size}{Numeric scaler for axis label size.}

\item{pt.size}{Numeric scalar for point size.}

\item{pt.color}{Character scalar for point color.}

\item{boxplot.fill}{Character scalar for boxplot fill color.}

\item{boxplot.line.color}{Character scalar for boxplot line color.}
}
\value{
plotly object
}
\description{
Plot selected information across lineages from DepMap.
}
\author{
Jared Andrews
}
